requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==4.0.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
phonenumbers==8.12.41
Pillow==9.0.0
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
tzdata==2021.5
